before this change, multiple calls to drmaa_set_attribute($jt, $DRMAA_NATIVE_SPECIFICATION, $drmaa_cmd); would overwrite the previous.
This change allows multiple calls to "stack".  Basically the strings are
appended together, using a space as the delimiter.

One side effect of this change is that the user must re-create the
template for each job. e.g., before this change a user could do this
- make template
- set native spec
- submit job
- change native spec
- submit job

but now, the native spec isn't changed, but it just appended to.  So the
second job wil contain both the first jobs native spec and its own
native spec.